make Eval templated so that autoreduce decisions are made systematically

+++ dali/tensor/tensor.cpp
@@ -298,13 +298,13 @@ Tensor Tensor::operator[](const Tensor& indices) const {
 
 
 Tensor Tensor::operator[](const std::vector<int>& indices) const {
-    Array indices_arr({indices.size()}, DTYPE_INT32);
+    Array indices_arr({(int)indices.size()}, DTYPE_INT32);
     indices_arr = indices;
     return tensor_ops::gather(*this, indices_arr);
 }
 
 Tensor Tensor::operator[](const std::initializer_list<int>& indices) const {
-    Array indices_arr({indices.size()}, DTYPE_INT32);
+    Array indices_arr({(int)indices.size()}, DTYPE_INT32);
     indices_arr = indices;
     return tensor_ops::gather(*this, indices_arr);
 }
